Add/modify cpu core settings for containers.
The container_settings.max_cpu_cores field contains the maximum
allowable CPU cores for a tool. It's a numeric field since that's what
both docker and kubernetes take for their cpu limits.

The container_settings.min_cpu_cores field was changed over to a numeric
field to mirror the container_settings.max_cpu_cores field. It's
currently unused, but could be used to assist job schedulers in HTCondor
or Kubernetes.