Various fixes based on feedback
Makes the idx_buckets_name index creation explicitly a unique index.
Adds IF NOT EXISTS to the index creation statements.
Added a dedicated CREATE UNIQUE INDEX for the (bucket_id, user_id) fields on the bucket_permissions table.
Removes the NOT NULL constraint on the bucket_permissions.created_by field and have it get set to NULL if the user is deleted.